DAP: Adapting the page input of one fragment or repeat to the mysql database inputs
One specific fragment or repeat that was selected in from the page
where all fragments or all repeats are displayed should be described
in detail. Inside this page for the fragment and repeat the information
should also be taken from the entries of the mysql database tables.
The structure of the rhtml files is adjusted to take the imputs from
the mysql database.