DAP: basic structure of the update page
The update page should be used to insert new fragments or
repeats or to change the database entries.
Here, the basic structure of this page is inserted.
Actually the page is visible for everyone, but in future
it should only be visible for administrators or people that
are permitted to insert new items or modify existing ones.
Additionally this includes a help page where the serveral
input formats are described.