TCoffee: Temporary deletion of some options
As the clustalw option for TCoffee does not work, as well as
the lower case option, they are deleted at the moment until they
are fixed.
Also, the tool commands in the t_coffee_action.rb are rewritten
in the usual manner as it is written for all other tools.