GENERAL: re-integrated ulimit
As it is important to have a limit for the file size
of the files (mainly the log files) that are written in
the tmp folder, the limit is set to 1000000.
Additionally, necessary because of the new submitting
system within the three tower cluster in Tuebingen, the
priority for qsub (-p) is set higher for the development
TK in Tuebingen.