DAP: rework of the rhtml pages and the display of the tool
To verify that only users with the dap group are allowed to modify
and to see the tool, the group dap was integrated and the tool
is only visible, if the user is logged in and if he belongs
to the group dap.
Additionally the webpages of the tool were rewritten.