Create new Release 2.9.0: Integration of Tool DAP
The tool DAP (Dictionary of Ancestral Peptides) is
integrated into the Toolkit at the section Classification.
At the moment, it is only for testing stage visible and
modifiable only for users within the group dap and
visible for users with the group admin. This tool includes
a collection of fragments and repeats.