Tool_controller: Labels of tools in forward box fixed.
As there was true instead of the tool label within the forward selection
box, the bug is fixed now and it is working again.
Additionally when selected the immediate queue, immediate is set to
TRUE.
Additionally the links (alignments and images for the DAP tool),
that are set in the false way are corrected now into the relative path.