DaTAA: inserted missing images and hydroplots
Within the sqlite3 database of dataa were images and
hydroplots in binary format. As something within the database
crashed, they could not be displayed anymore. Now, the
images and hydroplots are stored within the database folder
as .png files. Some minimal changes had to be done for
displaying the files within the Toolkit.