Quick2D: Replace wrong output of JNet
As for JNet was displayed a helix for nearly the whole sequence
which could not be correct, this line of the output (for which was
taken the FINAL line, was replaced by the ALIGN line (but could also
be the HMM line [just for the case, that the ALIGN line is not the
correct one ...]).

Second was to make sure that export and save results works again.