Repper: Insert new page for pasting own scale (FTwin)
Using Javascript and Ruby, a new page is inserted to make it
easier to insert own values for the scale of FTwin. Within this
template, the amino acid has to be inserted in one-letter-code,
followed by a blank and the number. The line has to be terminated
with a return.