Patch to avoid Mysql has gone away exception
Internal users logged in into the Toolkit often got the Mysql-has-gone-away exception
when submitting a job. External users only got displayed "500 Internal server error" instead.

The callstack in these situations contained the method ApplicationController.setup, inserted
as a before-filter into the execution sequence.
The error occured when trying to reload the session to get updated user info in a check action.

That exception is now caught in ApplicationController.setup. In case of that exception, the
connection is reestablished by calling ActiveRecord::Base.verify_active_connections! and the
reload call is executed again.

This only touches the symptomes, but why the connection goes away in these check actions still
is unknown. Why only check actions seem to get this exception is still unknown, too.