Adjust qsub command to work again on the internal cluster.
The h_rt request causes the job to be scheduled in the schort queue,
which has free resources most of the time. This results in shorter
waiting times.

The previous approach, to set the job priority with -p. does not
work any more on the internal cluster since the bidding system was
replaced in March 2013.