Tool HAMPpred: Keep compatibility with HHpred code
Tool HAMPpred is very similar to the tool HHpred. To be able to reuse code, the datatype of
corresponding variables should be the same in HHpred and HAMPpred.
Therefore, the actual databases list in HAMPpred is also stored as an array as is down with HHpred,
despite HAMPpred can only use its only special database.