PCoils tool doesn't finish execution
PCoils tool more and more often does not succeed, especially when option "Run PSIPRED" is turned on and
user is logged in.

It turned out, that blastpgp called by runpsipred.pl requires a little more then 4G of memory.
Therefore, for the pcoils tool, the memory limit is changed from 4G to 5G.
That's done generally for pcoils, because currently the value of the "Run PSIPRED" option is not
available easily in the configuration.