refactor jalview calls
Applet jalview did not open any more the jmol popup when clicking on the jalview button.

This seems special to the actual invocation of the unicorn webserver.
But to be able to modify the jalview usage and to implement a workaround,
a refactoring seemed appropriate:
The call of the jalview applet was factored out to a partial file.
To change the jalview call, now only one file has to be changed instead of about 35 files as before.

The current workaround is to specify the nojmol parameter.