Inform user about killed jobs
When a job is killed by the cluster system, the job display in the toolkit
keeps on showing "run" state and does not inform the user in any way.

The ruby sge_worker has been extended to install a signal handler.
The signal handler catches signals caused by s_rt and s_hvmem signals
and outputs appropriate error messages. Usage of s_rt and s_hvmem also
enables setting the lights to red.
Additionally, the notify flag is used. Now the signal handler can output
messages on the notify signals, too, and the lights can be switched to red
in case of some other operations (i.e. when someone kills a job with qdel).