Reset form button does not work as expected
After resubmitting a job (i.e. from an HHpred result), a new input window opens with the selected
sequence transferred to toe input sequence field. Clicking the reset form button in that situation
does not clear the sequence input field.

Technically, that is not a bug: The reset form button restores the initial contents of that window. That works.
The user might not see the difference: If he is viewing the input page, reset form clears the input fields. If he
is viewing the resubmit page (which looks the same but the header url), reset form restores the sequence.

To clarify the behaviour in the resubmit page, the labeling of the reset form button has been change (only) there
with "Restore input". A button labeled "New form" has been added to clear the input field. That button technically
behaves like a link to the index page. See FS#249.