Option -pe parallel <cpus>
The cluster engine didn't know about cpu/thread usage of toolkit jobs.
According to the administrator, that could result in using to much cpus and crashing.

Usage of threads still is to be reported by the cpus parameter to the toolkit model.
That information now is forwarded to the cluster using the option -pe parallel.