Specify time limits for sge engine
The toolkit specified the queues and the queues define time limits.
Because on some systems, the queues do not correspond directly to the toolkit queue classes,
the queue implementation should not be exploited in that way any more.

Now the default time limits can be specified in toolkit options (local_environment.rb).
The toolkit does not need to specify the real queue names any more.
To get optimal user experience, the time limits in local_environment.rb should consider
the queue setups: If no priority is required, the time limit can be omitted or be large.
If quick response time is wanted, the time limit should be in the range of an appropriate queue.

Currently, in some environment setups the real queue names still are used for compatibility.