PDBalert in most cases doesn't work
The logfiles showed an error. When PDBalert works for some input, clicking on the structure link showed an error page.

The memory limit for the grid engine had to be increased to 5G.
Now the modeller controller is used instead of the hhpred controller for the structure.

rsub was modified to be able to forward severel resource options.