Quick2D/Disopred2 exceeds memory limit
The Quick2D job of a toolkit user stopped in error state, because its memory limit was exxceeded. There also is a warning:
[blastpgp] WARNING:  [000.000]  posPurgeMatches: Due to purging near identical sequences, only the query is used to construct the position-specific score matrix

The blastgpg call in Disopred2 again needed more memory. The memory limit for Quick2D jobs has been increased from 22G to 25G.
The warning wasn't addressed. It still is given in the internal log files with the user's input.