listing hhblits databases
The database list offered by the tool hhblits was incomplete
and not ordered correctly.

A programming error has been removed which caused that the
database list was incomplete. As a consequence of that,
now the sortlist has to be unique (no sortlist element now
may be part of another sortlist element).

Now the database file list is sorted to make the ordering reproducable.
The relabeling feature for database names was changed to give more
control on the database names.

Be  cautious: The internal default database base name has to be uniprot20, because severel
tool models refer to that name. That database can be relabeled using the relabeling feature.