FS#230: Toolname as a job id is bad
A user used the Toolname psi_blast as a job id. Then, the tool could not be used any more: Clicking the submit button didn't start the tool.

There's already a validator for the job id, which checks, among other things, if it is a tool name.
Because of a side effect using gsub! the check for a tool name failed for tools with an undersore in the tool name, if the underscore
was given in the jobid, too. The use of gsub! therefore has been replaced by the use of gsub.