Rendering Internal Server Error causes exceptions
File production.log was cluttered with exception backtraces caused by problems when
trying to display "Internal Server Error". It was time consuming to find hints on
the real problem among this amount of text.

The exceptions were caused, because an original rails method for error handling has
been overloaded in 2006 to display a neutral page 500.html instead of mentioning a
rails problem. The rails interface seems to have changed since then, because the
overloading module called render with parameters not accepted any more.

The overloading is removed completely, because the original rails method
now seems to handle the "internal errors" in the way it was indented by the overloading.