Forward PsiBlast to PsiBlast by default
When forwarding results from PsiBlast, the default target tool was HHAlign. It should be another invocation of PsiBlast.
FS#361

The forwarding was changed. This didn't work immediatly, if a uniprot database has been used.
For that, the access mode list had to be synchronized with the tool list.
To be able to specify the default tool of the forwarding list by tool name instead of an index to the list of available tools,
a new public method named default_forwarding has been added to class ToolController.