match state assignment
Some users missed the possibility to specify a match state assignment.

This possibility now is added for the tools hhpred and hhsenser. It
already was available with tools hhblits. The appearance of the input field
has been brought in line among all these three tools and is based on
suggestions of Mr. Soeding.