FS#287 Selecting a user database is not intuitive
If a user selects a user database, the previously selected std database remains selected.
This is not intuitive, because the user may look at both choices as one interface element.

Now, if a user selects a user database or a standard database without holding the ctrl- or shift-key,
the selection in the other select control is unselected.