Statistics on average execution time of tools
Statistics on average execution time of tools are requested. See FS#372.
The purpose of that is to characterize the tools, not load testing.
Therefore, the user cpu time is used, not wallclock or sys time.

Base of the implementation is a time call on the execution of job scripts.
To relate these values to tools, the times for executing child jobs
and actions of the job are added all to the tool execution time.

Testing showed, that the times are mostly the same as the ru_utime value
given by qacct minus 2 seconds.

The significance of these values depends on the linux/bash time implementation:
The scheduling of processes can lead to exaggerated values, child processes might not be
included.