sge job names and variable cpu number
The names of cluster jobs for executing tools were prefixed with "TOOLKIT".
To facilitate associating a cluster job with a tool, that prefix now is replaced
by the tool shortcut.

When looking at the execution times of tools, it was detected, that msaprobs
in some computing environments allocated as much cpus as a node has. That prevents
the grit engine from distributing the ressources correctly.

Now sge_worker supports to some extend cpu range requests and is able to
provide a shell variable which contains the number of cpus allocated for
use in the job script. The msaprobs call has been changed to specify exactly
that number of cpus. For portability, that shell variable is also set by the worker
classes of the other computing environments, but, as now for simplicity, probably
not with the highest possible value. That should be improved, if those other
computing environments should be used in production.