misunderstandings with fasta validator
Relatively often, clustal format alignments are used as input.
Some tools accept only multiple sequences in fasta format.
The fasta validator then tells the user that there have to be at
least to sequences. That's because the fasta validator finds no
">gi" prefixes.

Now the special case of a clustal input is handled explicitely by
the fasta validator. It now tells the user that clustal alignments
are not allowed there.