reformat does not deal with html input
When forwarding from cs-, prot- or psi-blast to reformat, html format was used.
That's not useful, because reformat does not support html format.

Now alignment mode is used when forwarding from these tools to reformat.
This means, that alignhits_html.pl is used to create an alignment in fasta format.