Indicate in status.log, that the tool has finished
See FS#358.
Now, if for some reason, after the job is done the toolkit does not switch to the result page,
the user at least can recognize that the job is completed and that he can try
to manually invoke the result page (i.e. by replacing "waiting" by "results" in the browser url,
if s/he knows this trick).