redirect to :back error
Especially when using a proxy, "Internal Server Error" screens are caused by redirects to back.
Cause: The HTTP referer (the page where the request was submitted) was not known.

Redirect to back often is called, because an error was detected in the user input (in a form field)
and that page should be redisplayed highlighting that input error. Because most form fields are
in the index pages of the tools, now by heuristic the index page of the current tool is displayed
instead of the "Internal Server Error" screen by the input checking routine. Because that routine
already puts all the job parameters into the session, user information is not lost when this heuristic
was correct. If it's not, showing that index page still might be better than showing the "Internal
server error" page.

Redirect to :back also is called when checking the job status. Because with that no error messages
have been observed, that code wasn't changed.