Forwarding possibilities sometimes are wrong
The tools active for forwarding depend on the number of sequences checked for forwarding.
Counting the checked sequences was wrong sometimes, because it relies on
the fact, that exactly the same hits are selected in the hits section and
in the alignment section of the result page. With some tools, only
sequences in the hits section were checked by default.

Because the procedure of selecting hits by default is similar in several tools,
it first was factured out into a new module to be used by all of these tools.
That factured out procedure now should select hits as well in the hits section
as in the alignment section.