Killing jobs
If a user kills a job still waiting in the queue, the job gets into Eqw state, because the job file has been deleted,
but the job wasn't killed.

Now the qdel call has been reactivated and some checks on the job state were added or changed for error state.