Pasted results different from computed results
With some tools, a toolkit has the option to "paste an example sequence" and "paste result".
Pasting the example sequence and performing the tool most probably will lead to results different
from the pasted result.

The difference results from updated databases. Currently, no automatic update of these example results
is available. In princible, an automatic update of example results after updating an involved database
seems possible to implement.

Instead, near to the "paste result" button an explanation is displayed that mentions that the
pasted result probably is based on an older database version and therefore the pasted result
may be different from a newly computed result.
A specific pasted result may be updated by everybody manually by resubmitting the page and using the
reserved job name.