Tool HHpred crashes immediatedly after it starts
SgeWorker only computed a memory size of 15G.
Reason was, that I replaced action.type with action.class.
Object.type is synonym for Object.class, but Action inherits from ActiveRecord, which overloads the type method
by returning the class name instead of the class.
Switched back to that type method.