Psi-Blast job crashes or takes very long time
A Psi-Blast job with nr database, three iterations and input sequence
MRVGFYALLFASAALLASTAPLPNPPMTPAQHYAQAIHHEGLARHHTTVAEDHRQTANLHDNRIKAAKARYNAGLDPNGLTSAQKHQIERDHHLSLAAQAERHAATHNREAAYHRLHSQTPAPGTKRSIDELD
took 3 hours to complete, while it took only a few minutes on the Munich cluster.
When trying to reproduce that, the job crashed with exit status 134 (Aborted).

Without memory restrictions, everything was fine. Observing it with htop, then for short times nearly 38G of virtual memory was used.
The virtual memory request for PsiBlast now is set to 50G.

To avoid "Internal Server Error" pages only because of somewhow missing actions in the check procedure, Job.run_on now returns 0 instead
of throwing an exception when there are no actions.