Previously killed job reappears
After killing a job beeing in queued or running state, logging out and logging in,
the job reappeared in the job cart. Displaying it, leads to an error (displayed as
internal server error for external users). See FS#392.

Removing a job only deleted assiciated actions, workers, subjobs etc. from the
database, but kept the job itself. Therefore, the job reappeared.
Many toolkit pages try to access the first action of the job to get additional
information on the job (i.e. status, running time). But all of the actions
have been deleted.

To avoid this "bad" state of a job without actions, now when actions are removed
the job itself also gets removed from the database.

When the user from the browser history selects a previous page displaying an
already deleted job, a redirect_to :back was tried because the page of a
deleted job cannot be displayed any more. Then there's no known referrer to
redirect back. Now the toolkit should display the index page of the tool in
these cases.