Server error when displaying job list
If the task list of a user contains a job which doesn't appear in the .yml file of the tool,
a method on nil was called. That prevents the user from logging in.

Now that case is checked and logged. That job then is omitted from the job cart, so that the error does not occur.
This can happened with a new tool where not all jobs are configured in the yml-file yet.