sequence validation
Error messages given by the fasta validator seem to be formulated
in the assumption, that the input is fasta format.

It often happens, that the input is in another format than fasta and
the user didn't adjust the input format selection. Then the input may
be correct in another format and the error message given by the fasta
validator is misleading.

Now in some special cases, a hint to check the format selection is
appended to the error message or, if the other format is recognized,
the user is prompted to select that format.