maximum number of sequences in hhfilter
Tool HHfilter restricted the number of input sequences to 1000.
HHfilter is used to reduce the number of relevant sequences.
Users complaint, that HHfilter would be useful with a larger
number of input sequences.

The maximum number of input sequences was changed to 21000 in the
input validator specification of tool HHfilter.