users cannot log in to their account
From time to time, some user complains that she cannot log in to her account (see FS#402).

In these cases, it was sufficient to set the account's verified flag by directly using mysql. No
appropriate error messages were found in the log files. This commit only installs the usual
workarounds for the mysql server gone away error in class User.

If the error will still appear, the workflow for confirming the account has to be changed.
Probably a new confirmation E-Mail should be sent automatically, if a user tries to confirm
an account and the existing security token for confirming the account has expired.