Runtime errors because of database not usable
Tools patsearch and hmmer3 got into error state, if uniprot database was selected (FS#405).
That's because these tools don't accept pal files and uniprot is not a fasta file.

Now only databases usable by the tools are offered in the list of selectable databases.
This means, that uniprot is not offered there any more with these tools.
With tool patsearch the users should be able to find the obvious workaround: selected the trembl and swissprot databases.

With tool hmmer3, there is another issue:
Currently only one target parameter can be used. Therefore, currently multiple selection of databases and/or genomes has
been disabled.
The web page hmmer.janelia.org/search/hmmer does offer uniprot and multiple selection of proteomes, put at least the e-values
returned indicate, that that page only adds the results of separate calls of hmmer3 together. Because that's not
considered as correct, that approach wasn't followed.

Further improvement of the hmmer3 tool page wasn't done because of lack of ressources to do it.

This commit also applies further uses of the workaround for the "sql server has gone away" error.