use nr70 as default database with blast tools
For the bioinformatic course, a small database is preferable as the default database.
That hopefully will reduce the possibility of cluster overloading.

Now with blast tools (only), the ordering of the standard databases is changed to
show nr70 at the top position. Hopefully that prevents additionally from selecting the
nr database.

Shared template protein_dbs.rhtml was changed to select the top database as the default
database instead of always selecting nr.

As a side effect, changing the database was removed from the "Paste example sequence" function.
Now this function does only what it pretends to do.