Forward HMM often does not work
The result pages of some tools offered the function "Forward HMM". The intent was, that
the tool forwarded to can use the result alignment and does not have to recompute it.

In some cases, that didn't work, because some tool offering "Forward HMM" didn't produce an
HHM at all.
That wasn't fixed by creating the HMM, because the receiving tool in most cases won't use the
HMM file, but would only use the A3M file also forwarded by "Forward HMM".
Because of that, "Forward HMM" was generally replaced by "Forward MSA", which uses the
A3M file of the result only. If the receiving tool needs an HMM file, it can recreate it from
the A3M file without much overhead.

"Forward MSA" now in general offers the tools hhblits, hhpred, hhrep and hhrepid as receiving tools.

The user interface was kept largely. That means, that the forwarded A3M file isn't shown to
the user for editing. With a forwared HMM, that could be justified in that the hhm file is binary.
Now, it's expected that users may wish to be able to view and modify the MSA, before they submit
the receiver tool job.
That's not implemented by now. To do that, it has to be discussed, how to make it clear to the user
that this forwarded MSA (in form of an A3M file) is handled differently than a standard alignment
input in a3m format: For the latter, many tools still invoke psiblast or hhblits to create a MSA,
while this has to be omitted for the forwarded MSA.