Hit numbering
It was suggested, that the presentation of blast output should be improved by numbering the hits.
It was mentioned, that on the NCBI page, a section "Sequences not found previously ..." is used to
show changes from the last iteration.

Tools considered relevant to these suggestions: ProtBlast, PSI-BLAST and CS-BLAST.

Looking at the toolkit implementation, at least tool PSI-BLAST tries also to provide a section
"Sequences not found previously...", but includes only sequences below threshold in that section.
Other sequences are appended to the "worse hits" section. For unification of presenting blast
output, it still has to be discussed, if that is usable and if it also would be benificial with other tools.

The hit entries now are numbered in tools ProtBlast, PSI-BLAST and CS-BLAST.
This change only affects the presented view, which has been unified for these three tools.
That also required the implementation of a corresponding, common interface of the models of these three tools.
To provide the "not found previously" section with other tools than PSI-BLAST, the tool models would have to
be further unified.