MySQL server lost connection error with insert
In case of a lost connection error, it is not known if the insert was performed or not.
If it was performed, it may not be called once more, because i.e. a spare action
might prohibit a job from getting into state done.

Now, the workaround with save actions were refactored to module ProtectedSql.
There, errors in insert operations are examined.