uploaded database causes errors
After uploading a user database, selecting it and submitting the job, the job gets into error state.

By uploading, the new database is also formatted etc. Then the database is added to the personal database list
in the user interface. While doing that, a damaged file name was used. Therefore, the database wasn't found by the job.
That does not happen after reloading the tool page. Then the correct file name loaded from the MySQL database is used.

Now the damaging of the file name is avoided.