User module issues
1. After a tool is renamed (more formal: if a job class is renamed) and there are still user jobs refering that name,
the user cannot log in, because these jobs cause errors.

This does not happen in normal operation. Therefore, these errors only are caught to enable the user to log in.
His job list will then not be displayed until the jobs are deleted from the database (by expiration or manually).

This is mainly an issue in development versions of the toolkit.
In the productive version, classes implementing a tool only should be removed, after all corresponding jobs, actions etc. have been
removed from the MySQL database.

2. In user.rb, there's some code which should "allow logins for deleted accounts". It wasn't checked in which circumstances this code
would be invoked. But that code would try to set attribute "authenticated_by_token" to true in the MySQL database.
Because that attribute doesn't exist there and no usage of it was detected, that statement was commented out.