Prot Blast should use 2 cpus
In the friday meeting, it was decided, that tool Protein BLAST should use more than one cpu
in its blast call.

The default number of cpus has been changed to 2, but with the correct advanced options,
one can request any number of cpus now.
The default number only is two, because in the bioinformatics course, a lot of prot blast jobs
are started usually.