cleanup blast output processing
PsiBlast is the first tool, for which a Blast+ version is implemented.
To be able to use that approach for other tools like ProtBlast,
the code was cleaned up:
External interfaces (i.e. to the evalue-threshold values) should be unified,
Only locally used variables (like the hits' key list) were removed from the external interface.