using ruby1.8.7 with ubuntu 14
For using ruby 1.8.7 with ubuntu 14, the ruby version manager (rvm) can be used.
rvm is activated in login shells, where also can be switched already to ruby 1.8.7.
This didn't work with the qupdate mechanism in the toolkit.

The reason was, that qupdate was called with an extra shell which restored the original environment
without rvm and ruby 1.8.7. That can be avoided by not creating an extra shell.

The bioinformatics script still is called in an extra shell. Because rails shouldn't be used in
those scripts, it should be possible to write ruby code which works with ruby 1.8.7 as well as with
more recent ruby versions.