hhsenser crashed because of queue time limit
The tmax parameter used internally for limiting the computing time should prevent, that
the queue time limits gets exceeded.

hhsenser stops looping when tmax is reached. Then it summarizes the results.
The loop body also takes time, so the running time already can be 10 minutes above the time limit
for looping when checking it. Summarizing the results also takes time.

Therefore, tmax was decreased from 15 to 25 minutes below the maximum running time in the cluster for toolkit jobs.
tmax is specified by a constant, not derived directly from the maximum running time in the cluster.
Therefore, if the maximum running time in the cluster changes, tmax has to be adjusted in
hhsenser_action.rb