Internal error while updating status
An internal error happend while updating the status of a job.
The job task didn't require submitting a queue job.
The time is computed by an aggregate function on all queue jobs.

In rails 2.0.2, if there are no columns, an aggregate function always
returns nil. Later that was changed, i.e. with more current rails versions,
the aggregate function "sum" returns 0 if there's no column.

QueueJob now returns total time 0, when there are no workers, so this time
now always can be added to the time of an action.