Forwarding displays the wrong tools
When the results depend on a uniprot database, tool blammer is removed from the
list of tools to be forwarded to after computing the list of suitable tools.
With ProtBLAST and CS-BLAST, the indexes of suitable tools was not adjusted,
therefore the wrong tools were enabled for forwarding.

The solution already implemented with tool PsiBlast last year was copied.