Sorting of Jobs in Result slider
The ording of jobs in the sidebar section of the toolkit wasn't comprehensible.
In fact, it was according to the internal database representation.

I found, that in 2011 already a change was made to sort the jobs on creation date.
This change wasn't effective because of a misspelling of the column name.

Now, jobs are sorted according to the creation date.
If one clicks on the sorting arrows, that column will be the first criteria for sorting,
the second will be the creation date. This sorting immediately gets lost when the
page is redisplayed, because neither the newly sorted list of jobs nor the current
sorting criteria is kept anywhere after using it once.