Wild card databases not available
The pattern "\w+" for selecting any availabe database does not work, because this immediately is reduced to "w+".
The Backslash has to be doubled.

Then, in a loop to collect the available databases, the collected ones are removed from the available databases.
That's to avoid some databases beeing matched and collected several times.
It doesn't work removing members of an array just beeing iterated over - some databases get lost.
The iteration now is based on a copy of the database array.