export of hhpred result doesn't work
From several tab pages of the result page of hhpred, export/save didn't work.

It seemed, that the code was copied from hhblits, but the result files of hhpred have different names.

Now the export/save actions of the tabs reducedalign, representativealign and fullalign export/save
what also is in the text box of the result page.

It also has been detected, that the links "Create model" and "Forward to PCoils" were inconsistent:
When there was no hit from PDB or SCOP, the user didn't get a hint why he could not create a model.
If he switched from "Create model" to "Forward to PCoils", the checkboxes remained disabled.