Crash when viewing the results of HHpred
An unexpected query header could cause a crash, i.e.
>4983_02905
MND...

There, 4983 was taken as a pdb identifier. Then a chain id was searched in front of the underscore.
The code already tried to allow missing chain ids, but then a nil reference occured.

Now, this header still is treated as a possibly yet unknown pdb identifier, but doesn't crash the system any more.