Problems with forgotten password (FS#424)
There was an error when trying to leave the change password dialog.

When the change password dialog is visible, the user is not logged in.
(If he wants to change the password while beeing logged in, he uses the edit user dialog instead.)

The old "Cancel" action took the user to the welcome dialog, which requires, that the user is logged in.
Probably it was the intention, that the user is logged in automatically after clicking the "change password" link in his E-Mail,
but that's not implemented. Logging in automatically probably is not a good idea.

The "Cancel" link was replaced by a "Login" link, which takes the user to the log in page.