Commented out the code for displaying the info banner
Due to a network issue, the cluster system was down for two
days (16.07.16 and 17.07.16). We used a banner to inform our
users about this issue. The Toolkit is functioning normally
now and therefore the code for displaying the banner was
commented out. This code could be used for announcements
in the future.